# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "2.0"
  title: "Standard Interfejsu Rejestrów Akcjonariuszy"
  license:
    name: "Licence -> Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v1"
schemes:
- "https"
paths:
...
  /getAllOffers:
    post:
      tags:
      - "Offers"
      summary: "All offers from TOK"
      operationId: "getAllOffersPOST"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "In requests head should be attached token from login service"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/AllOfferEnvelope"
      security:
      - X-Auth-Token: []
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/OfferEnvelope"
        "400":
          description: "Wrong body"
          schema:
            $ref: "#/definitions/MessageResponse"
        "403":
          description: "Authorization error"
          schema:
            $ref: "#/definitions/MessageResponse"
        "404":
          description: "Offer not found"
          schema:
            $ref: "#/definitions/MessageResponse"
        "500":
          description: "Server error"
          schema:
            $ref: "#/definitions/MessageResponse"
      x-swagger-router-controller: "Offers"
  /getOffers:
    post:
      tags:
      - "Offers"
      summary: "All offers for specific issuer"
      operationId: "getOffersPOST"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "In requests head should be attached token from login service"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/OfferForIssuerEnvelope"
      security:
      - X-Auth-Token: []
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/OfferEnvelope"
        "400":
          description: "Wrong body"
          schema:
            $ref: "#/definitions/MessageResponse"
        "403":
          description: "Authorization error"
          schema:
            $ref: "#/definitions/MessageResponse"
        "404":
          description: "Offer not found"
          schema:
            $ref: "#/definitions/MessageResponse"
        "500":
          description: "Server error"
          schema:
            $ref: "#/definitions/MessageResponse"
      x-swagger-router-controller: "Offers"
  /getOffer:
    post:
      tags:
      - "Offers"
      summary: "Details for specific offer"
      operationId: "getOfferPOST"
      consumes:
      - "application/json;charset=utf-8"
      produces:
      - "application/json;charset=utf-8"
      parameters:
      - name: "X-Auth-Token"
        in: "header"
        description: "In requests head should be attached token from login service"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/OfferDetailsEnvelope"
      security:
      - X-Auth-Token: []
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/OfferEnvelope"
        "400":
          description: "Wrong body"
          schema:
            $ref: "#/definitions/MessageResponse"
        "403":
          description: "Authorization error"
          schema:
            $ref: "#/definitions/MessageResponse"
        "404":
          description: "Offer not found"
          schema:
            $ref: "#/definitions/MessageResponse"
        "500":
          description: "Server error"
          schema:
            $ref: "#/definitions/MessageResponse"
      x-swagger-router-controller: "Offers"
...
securityDefinitions:
  X-Auth-Token:
    description: "Token for login service, Token validity period 10 second, each request\
      \ extends the validity of the token."
    type: "apiKey"
    name: "X-Auth-Token"
    in: "header"
definitions:
...
  PageEnvelope:
    type: "object"
    example:
      page: 1
      items_on_page: 20
  PageEnvelopeResponse:
    type: "object"
    example:
      page: 1
      items_on_page: 20
      total_pages: 20
...
  AllOfferEnvelope:
    type: "object"
    required:
    - "page_envelope"
    properties:
      page_envelope:
        $ref: "#/definitions/PageEnvelope"
    example:
      page_envelope:
        page: 1
        page_total: 1
  OfferEnvelope:
    type: "object"
    required:
    - "offers"
    - "page_envelope"
    properties:
      offers:
        type: "array"
        items:
          $ref: "#/definitions/OfferMetadata"
      page_envelope:
        $ref: "#/definitions/PageEnvelopeResponse"
    example:
      offers:
      - order_id: 12345
        krs_e_number: "2313445"
        status: "B"
        negotiations:
        - paper_count: 12
          price: "21"
          negotiation_id: "12"
      page_envelope:
        page: 1
        items_on_page: 2
        total_pages: 1
...
  OfferMetadata:
    type: "object"
    required:
    - "order_id"
    properties:
      order_id:
        type: "integer"
      krs_e_number:
        type: "string"
      status:
        type: "string"
      negotiations:
        type: "array"
        items:
          $ref: "#/definitions/NegotiationEnvelope"
    example:
      order_id: 12345
      krs_e_number: "2313445"
      status: "B"
  NegotiationEnvelope:
    type: "object"
    required:
    - "data"
    properties:
      data:
        $ref: "#/definitions/Negotiation"
    example:
      data:
        paper_type: "nonfungible stocks"
        series: "AZSA"
        paper_number_from: 1221
        paper_number_to: 2345
  Negotiation:
    type: "object"
    required:
    - "negotiation_id"
    properties:
      paper_count:
        type: "integer"
      price:
        type: "string"
      negotiation_id:
        type: "string"
    example:
      paper_count: 200
      price: "12"
      negotiation_id: "12"
  OfferForIssuerEnvelope:
    type: "object"
    required:
    - "krs_e_number"
    - "page_envelope"
    properties:
      krs_e_number:
        type: "string"
      page_envelope:
        $ref: "#/definitions/PageEnvelope"
    example:
      krs_e_number: "121321"
      page_envelope:
        page: 1
        items_on_page: 20
  OfferDetailsEnvelope:
    type: "object"
    required:
    - "page_envelope"
    properties:
      krs_e_number:
        type: "string"
      page_envelope:
        $ref: "#/definitions/PageEnvelope"
    example:
      order_id: 121321
      page_envelope:
        page: 1
        items_on_page: 20
...
  MessageResponse:
    type: "object"
    required:
    - "comment"
    properties:
      comment:
        type: "array"
        items:
          type: "string"
    description: "Status of new offer."
...